remove the pause button
I only ever introduced this due to performance/energy concerns on
low-powered devices, but we're doing fine there and the whole point of
the demo is the animation. The old pause behavior was also not quite
right, as unpausing would not resume the animation at the same point,
but it would change based on the elapsed time, which isn't right.