prevent suble layout shifts due to resize timeouts
Rapidly tapping on a mobile device in the hero section can trigger
a resize event, even if it looks like nothing changed due to minor
temporary zooms or other gestures. This tries to validate if anything
is actually different from before when the timeout finishes, rather than
blindly resetting the network.